+++ J4JMapLibrary/config/ISourceConfiguration.cs
@@ -6,8 +6,8 @@ public interface ISourceConfiguration
     int MaxRequestLatency { get; }
     string Copyright { get; }
     Uri? CopyrightUri { get; }
-    double MaxLatitude { get; }
-    double MinLatitude { get; }
-    double MaxLongitude { get; }
-    double MinLongitude { get; }
+    float MaxLatitude { get; }
+    float MinLatitude { get; }
+    float MaxLongitude { get; }
+    float MinLongitude { get; }
 }
